moving todomvc to ts
