+++ examples/package.json
@@ -8,17 +8,21 @@
     "counter": "browserify -p tsify counter/app.jsx | uglifyjs > counter/app.js",
     "frp-counter": "browserify -p tsify frp-counter/app.jsx | uglifyjs > frp-counter/app.js",
     "type-n-search": "browserify -p tsify type-n-search/app.jsx | uglifyjs > type-n-search/app.js",
-    "alacarte": "browserify -p tsify alacarte/app.jsx | uglifyjs > alacarte/app.js"
+    "alacarte": "browserify -p tsify alacarte/app.jsx | uglifyjs > alacarte/app.js",
+    "todomvc": "browserify -p tsify todomvc/app.tsx | uglifyjs > todomvc/app.js"
   },
   "dependencies": {
     "@reactivex/rxjs": "^5.4.0",
+    "classnames": "^2.2.5",
     "react": "^15.5.4",
     "react-dom": "^15.5.4",
     "xreact": "latest"
   },
   "devDependencies": {
+    "@types/react-dom": "^15.5.0",
     "alacarte.js": "^0.2.0",
     "browserify": "^14.3.0",
+    "jest": "^20.0.4",
     "ramda": "^0.23.0",
     "rest": "^2.0.0",
     "tsify": "^3.0.1",
