--- examples/quarantine.d/todomvc/README.md
@@ -1,7 +1,0 @@
-https://reactive-react.github.io/react-most/examples/todomvc/public/
-
-```
-npm install
-npm run build
-npm start
-```
