+++ examples/todomvc/components/interfaces.ts
@@ -1,0 +1,9 @@
+export interface Todo {
+  id: number
+  text: string
+  done: boolean
+}
+export interface MainSectionProps {
+  todos: Todo[]
+  filter: (todos: Todo[]) => Todo[]
+}
