+++ examples/todomvc/index.html
@@ -1,0 +1,19 @@
+<!doctype html>
+<html lang="en">
+	<head>
+		<meta charset="utf-8">
+		<meta name="viewport" content="width=device-width, initial-scale=1">
+		<title>Template • TodoMVC</title>
+    <link rel="stylesheet" href="//rawgit.com/tastejs/todomvc-common/master/base.css">
+		<link rel="stylesheet" href="//rawgit.com/tastejs/todomvc-app-css/master/index.css">
+	</head>
+	<body>
+    <section id="app" class="todoapp"></section>
+		<footer class="info">
+			<p>Double-click to edit a todo</p>
+			<p>Created by <a href="https://github.com/jcouyang">Jichao Ouyang</a></p>
+			<p>Part of <a href="http://todomvc.com">TodoMVC</a></p>
+		</footer>
+		<script src="app.js"></script>
+	</body>
+</html>
