+++ examples/todomvc/package.json
@@ -1,0 +1,70 @@
+{
+  "name": "xreact-todomvc",
+  "version": "0.0.1",
+  "description": "",
+  "browserify": {
+    "transform": [
+      [
+        "babelify",
+        {
+          "extensions": [
+            ".es6",
+            ".jsx",
+            ".js"
+          ]
+        }
+      ],
+      "envify"
+    ]
+  },
+  "scripts": {
+    "build": "NODE_ENV=production browserify src/app.jsx --extension=.jsx --extension=.es6 | java -jar bin/compiler.jar > public/app.js",
+    "start": "ecstatic -p 8000 public",
+    "watch": "watchify -d src/app.jsx --extension=.jsx --extension=.es6 -o public/app.js -dv",
+    "test": "jest --coverage"
+  },
+  "jest": {
+    "moduleFileExtensions": [
+      "js",
+      "jsx",
+      "es6"
+    ]
+  },
+  "dependencies": {
+    "classnames": "^2.2.0",
+    "most": "^1.0.5",
+    "ramda": "^0.22.1",
+    "react": "^15.3.2",
+    "react-dom": "^15.3.2",
+    "react-most": "^1.1.0",
+    "rest": "^1.3.1",
+    "union-type": "^0.3.3"
+  },
+  "devDependencies": {
+    "@types/ramda": "^0.0.9",
+    "babel": "^6.5.2",
+    "babel-jest": "^16.0.0",
+    "babel-plugin-ramda": "^1.1.6",
+    "babel-plugin-transform-react-jsx": "^6.1.18",
+    "babel-preset-es2015": "^6.1.18",
+    "babelify": "^7.2.0",
+    "browserify": "^12.0.1",
+    "ecstatic": "^1.3.1",
+    "envify": "^3.4.1",
+    "jest-cli": "^0.7.0",
+    "react-addons-test-utils": "^15.3.2",
+    "react-most-spec": "^0.2.3",
+    "watchify": "^3.6.1"
+  },
+  "author": "Jichao Ouyang",
+  "license": "ISC",
+  "babel": {
+    "presets": [
+      "es2015"
+    ],
+    "plugins": [
+      "transform-react-jsx",
+      "ramda"
+    ]
+  }
+}
