+++ examples/todomvc/todos.json
@@ -1,0 +1,9 @@
+[{
+  "text": "Try React Most",
+  "done": true,
+  "id": 0
+},{
+  "text": "Give it a Star on Github",
+  "done": false,
+  "id": 1
+}]
